fix: refactors TopNav into multiple files
